massive refactors to documentation generator script
